Made a mywork folder and more
I stored all the webpages but login.aspx to the mywork folder. I also created an experimental login page
