+++ Assignment_4_GC/Login.aspx.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Web;
+using System.Web.Security;
 using System.Web.UI;
 using System.Web.UI.WebControls;
 
@@ -83,6 +84,64 @@ namespace Assignment_4_GC
 
 
              
+        }
+
+        protected void Login1_Authenticate(object sender, AuthenticateEventArgs e)
+        {
+            //access the dbml
+            dbcon = new KarateSchoolDataContext(connString);
+
+
+            try
+            {
+
+
+                //Query to match the UserName and Password with something from the NetUser tables
+                var selectedUser = (from x in dbcon.NetUsers
+                                    where x.UserName == Login1.UserName && x.UserPassword == Login1.Password
+                                    select x).First();
+
+                //If password and username dont match, then show an error
+                if (selectedUser == null)
+                {
+                    //Show the error to the user
+                    Label6.Text = "Invalid UserName or Password";
+                }
+                else
+                {
+                    //Takes the userType from the selectedUser and the userID from the selected user and set them to variables
+                    string userType = selectedUser.UserType.ToString().ToLower();
+                    UserDetails.userID = selectedUser.UserID;
+
+                    //Adds the user ID, so we can grab data from the database after the webpage changes
+                    Session.Add("UserID", UserDetails.userID);
+
+                    //This determines which webpage the user will be taken to
+                    switch (userType)
+                    {
+                        case "administrator":
+                            Response.Redirect("~/mywork/Administrator.aspx", false);
+                            break;
+                        case "member":
+                            Response.Redirect("~/mywork/Member.aspx", false);
+                            break;
+                        case "instructor":
+                            Response.Redirect("~/mywork/Instructor.aspx", false);
+                            break;
+
+                    }
+
+
+
+                }
+            }
+            catch (Exception ex)
+            {
+                
+
+                //Show the error to the user
+                Label6.Text = ("Exception: " + ex.Message);
+            }
         }
     }
 }
