+++ Assignment_4_GC/bin/Assignment_4_GC.dll.config
@@ -25,6 +25,13 @@
         <add assembly="Microsoft.AspNet.Web.Optimization.WebForms" namespace="Microsoft.AspNet.Web.Optimization.WebForms" tagPrefix="webopt" />
       </controls>
     </pages>
+
+	  <authentication mode="Forms">
+		  <forms name=".ASPXFORMSDEMO" loginUrl="logon.aspx"
+			  protection="All" path="/" timeout="30" />
+	  </authentication>
+
+
   </system.web>
   <runtime>
     <assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
