+++ Assignment_4_GC/mywork/Member.aspx
@@ -1,0 +1,48 @@
+﻿<%@ Page Title="" Language="C#" MasterPageFile="~/MasterPage.Master" AutoEventWireup="true" CodeBehind="Member.aspx.cs" Inherits="Assignment_4_GC.Member" %>
+<asp:Content ID="Content1" ContentPlaceHolderID="head" runat="server">
+    <style type="text/css">
+        .auto-style2 {
+            width: 100%;
+            height: 142px;
+        }
+        .auto-style3 {
+            height: 24px;
+        }
+        .auto-style4 {
+            background-color: #99CCFF;
+        }
+    </style>
+</asp:Content>
+<asp:Content ID="Content2" ContentPlaceHolderID="ContentPlaceHolder1" runat="server">
+    <h1 class="auto-style4"><strong>Member Page</strong></h1>
+    <h1>Welcome
+        <asp:Label ID="lblMemberName" runat="server" Text="Member Name"></asp:Label>
+    </h1>
+    <p>
+        <table class="auto-style2">
+            <tr>
+                <td class="auto-style3">Member Details:</td>
+            </tr>
+            <tr>
+                <td>
+                    <asp:GridView ID="GridView1" runat="server" BorderStyle="Solid" CellPadding="4" ForeColor="#333333" GridLines="None">
+                        <AlternatingRowStyle BackColor="White" />
+                        <EditRowStyle BackColor="#2461BF" />
+                        <FooterStyle BackColor="#507CD1" Font-Bold="True" ForeColor="White" />
+                        <HeaderStyle BackColor="#507CD1" Font-Bold="True" ForeColor="White" />
+                        <PagerStyle BackColor="#2461BF" ForeColor="White" HorizontalAlign="Center" />
+                        <RowStyle BackColor="#EFF3FB" />
+                        <SelectedRowStyle BackColor="#D1DDF1" Font-Bold="True" ForeColor="#333333" />
+                        <SortedAscendingCellStyle BackColor="#F5F7FB" />
+                        <SortedAscendingHeaderStyle BackColor="#6D95E1" />
+                        <SortedDescendingCellStyle BackColor="#E9EBEF" />
+                        <SortedDescendingHeaderStyle BackColor="#4870BE" />
+                    </asp:GridView>
+                </td>
+            </tr>
+        </table>
+    </p>
+    <h4>&nbsp;</h4>
+    <p>
+        &nbsp;</p>
+</asp:Content>
