+++ Assignment_4_GC/mywork/Web.config
@@ -1,0 +1,11 @@
+﻿<?xml version="1.0"?>
+<configuration>
+  <system.web>
+
+	  <authorization>
+		  <deny users="?" />
+	  </authorization>
+
+
+  </system.web>
+</configuration>
